feat: #1132 PR 리뷰 반영
