+++ src/main/kotlin/study/racinggame/ui/RacingGameResultPrinter.kt
@@ -5,6 +5,6 @@ import study.racinggame.domain.RacingGameResult
 object RacingGameResultPrinter {
     fun print(racingGameResult: RacingGameResult) {
         racingGameResult.racingGameStages.forEach { CommandLineRacingGameStagePrinter.print(it) }
-        println("${racingGameResult.winners.joinToString(separator = ", ")}가 최종 우승했습니다.")
+        println("${racingGameResult.nameOfWinners().joinToString(separator = ", ")}가 최종 우승했습니다.")
     }
 }
